Updated packets and form title.
Updated watchdog and NRST messages

+++ AAUSAT6-OBC/Core/Src/DCMotorDriver.h
@@ -16,13 +16,13 @@
 
 void motor_init(TIM_HandleTypeDef *htimer, uint32_t channel);
 
-void motor_setPwm(uint8_t dutycycle);
+void motor_setPwm(float dutycycle);
 
 void motor_setSpeed(uint8_t speed);
 
-uint8_t motor_speedToDutycycle(uint8_t speed);
+float motor_speedToDutycycle(uint8_t speed);
 
-void motor_start(int8_t speed, char dir);
+void motor_start(uint8_t speed, char dir);
 
 void motor_stop();
